+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [TD-3753] Build using Elixir 1.12 and Erlang/OTP 24
+
 ### Added
 
 - [TD-3503]
-  - Created APIs `api/users/search` and `api/groups/search` 
+  - Created APIs `api/users/search` and `api/groups/search`
   - Removed permissions for non-admin users to `:index` users and groups
-  
+
 ## [4.20.0] 2021-05-17
 
 ### Added