+++ CHANGELOG.md
@@ -5,12 +5,16 @@
 ### Changed
 - [TD-3102] Improve change password functionality
 
+### Changed
+
+- [TD-3753] Build using Elixir 1.12 and Erlang/OTP 24
+
 ### Added
 
 - [TD-3503]
-  - Created APIs `api/users/search` and `api/groups/search` 
+  - Created APIs `api/users/search` and `api/groups/search`
   - Removed permissions for non-admin users to `:index` users and groups
-  
+
 ## [4.20.0] 2021-05-17
 
 ### Added